Fix failing feature test
The Submit feedback scenario was broken in master.
This caused the build not passing in CI due to the failing scenario.

The user flow for testing this scenario was loading twice the feedback
path before submitting the feedback. That was causing the
"navigate back" button was redirecting to the feedback submission page
as was set as the double load of the feedbacks submissions page was
setting it as the request referral, instead of the original rootpage
as intended.